Move JeProfService.h file into JeProfService.cc file
